reorganizes folder and file structure for back-end modules to be seperated by router in which they are used
