+++ server/routes/request.router.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const pool = require('../modules/pool');
 const router = express.Router();
-const craftTable = require('../modules/craftTable');
+const craftTable = require('../modules/router-modules/request-router/craftTable');
 
 /**
  * GET route template
